pyproject.toml
[project]
name = "guesty-microservice"
version = "0.1.0"
description = "FastAPI microservice with DDD and SOLID"
authors = ["Yolanda Mauri <ymauri@gmail.com>"]
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0",
    "python-jose>=3.3.0",
    "passlib[bcrypt]>=1.7.4"
]

[tool.setuptools.packages.find]
where = ["app"]
